chore: extract per-domain types.ts and lift this-free helpers out of auth-client
